requirements.txt
asgiref==3.3.1
autopep8==1.5.4
coverage==5.3.1
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
pycodestyle==2.6.0
PyJWT==2.0.0
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
